renamed and moved config/config.js to js/user-data.js & fixed how parameters are passed to Object.assign()
